requirements.txt
Flask==2.3.3
flask-basicauth==0.2.0
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
